made the following changes:
fixed a bug where messages which have errors ended up showing a new window
(of 0 size) on the MessagePanel.  fixed this by changing it so that
createMessageUI() can now throw a MessagingExcepion, in which case the action
itself is what shows the error.

also added a showError(String message, Exception e) to the ErrorHandler and
all files which implement it.

fixed two bugs:  one where a message is improperly cached.  now improperly
cached messages will throw an exception and then uncache themselves.  second
was that getSize() was trying a getContent(), meaning that, if the encoding
type for the message were X-UNKNOWN (or if any other exception like that
happens that we have ways of dealing with), then the request would fail.