messages from sent folders should show up properly in search folders now.
also, empty values should now show up as blank rather than as
'loading...'.

if autoCache is set, you now connect to the folder and show the
messages before you start to download the messages to cache them.

changed the MailServer code so that it exits out properly when you
shut down.